pyproject.toml
[project]
name = "finevision-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.4",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
]
